+++ custom_components/imou/icons.json
