+++ custom_components/imou/manifest.json
@@ -2,13 +2,13 @@
   "domain": "imou_life",
   "name": "Imou Life",
   "version": "0.0.1",
-  "documentation": "https://www.home-assistant.io/integrations/imou",
+  "documentation": "https://www.home-assistant.io/integrations/imou_life",
   "dependencies": [],
   "requirements": [
-    "pyimouapi==1.0.3"
+    "pyimouapi==1.0.4"
   ],
   "codeowners": [
-    "@ImouLife"
+    "@Imou-OpenPlatform"
   ],
   "config_flow": true,
   "iot_class": "cloud_polling"
