+++ custom_components/imou/select.py
@@ -6,7 +6,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .base import ImouEntity
+from .entity import ImouEntity
 from .const import DOMAIN, PARAM_OPTIONS, PARAM_CURRENT_OPTION
 from pyimouapi.exceptions import ImouException
