+++ custom_components/imou/sensor.py
@@ -4,7 +4,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .base import ImouEntity
+from .entity import ImouEntity
 from .const import DOMAIN
 
 _LOGGER: logging.Logger = logging.getLogger(__package__)
